fixed int to float implicit conversion warning for the INT_MAX boundary check

+++ m06/ex00/ScalarConverter.cpp
@@ -6,7 +6,7 @@
 /*   By: razaccar <razaccar@student.42lausanne.ch>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/29 18:03:10 by razaccar          #+#    #+#             */
-/*   Updated: 2025/09/30 15:46:37 by razaccar         ###   ########.fr       */
+/*   Updated: 2025/11/11 19:15:38 by razaccar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -100,7 +100,7 @@ void	ScalarConverter::convertFloat(float f) {
 	else if (isprint(static_cast<int>(f)) == false) std::cout << "Non displayable" << std::endl;
 	else std::cout << "'" << static_cast<char>(f) << "'" << std::endl;
 	std::cout << "int: ";
-	if (f < INT_MIN || f > INT_MAX) std::cout << "Out of bounds" << std::endl;
+	if (f < INT_MIN || f > (float)INT_MAX) std::cout << "Out of bounds" << std::endl;
 	else std::cout << static_cast<int>(f) << std::endl;
 	std::cout << "float: " << f << "f" << std::endl;
 	std::cout << "double: " << static_cast<double>(f) << std::endl;
@@ -112,10 +112,10 @@ void	ScalarConverter::convertDouble(double d) {
 	else if (isprint(static_cast<float>(d)) == false) std::cout << "Non displayable" << std::endl;
 	else std::cout << "'" << static_cast<char>(d) << "'" << std::endl;
 	std::cout << "int: ";
-	if (d < INT_MIN || d > INT_MAX) std::cout << "Out of bounds" << std::endl;
+	if (d < INT_MIN || d > (double)INT_MAX) std::cout << "Out of bounds" << std::endl;
 	else std::cout << static_cast<int>(d) << std::endl;
 	std::cout << "float: ";
-	if (d < FLT_MIN || d > FLT_MAX) std::cout << "Out of bounds" << std::endl;
+	if (d < FLT_MIN || d > (double)FLT_MAX) std::cout << "Out of bounds" << std::endl;
 	else std::cout << static_cast<float>(d) << std::endl;
 	std::cout << "double: " << d << std::endl;
 }
